move around some stuff and added initial folders
